move cache to module
